refactor: put modal and table into component
